+++ src/utils/index.js
@@ -1,4 +1,5 @@
 export { default as AppRoute } from 'utils/AppRoute';
+export { default as LanguageCheck } from 'utils/LanguageCheck';
 export { default as MainRouteRedirect } from 'utils/MainRouteRedirect';
 export { default as RoleChecker } from 'utils/RoleChecker';
 export { default as RoutesSwitch } from 'utils/RoutesSwitch';